+++ dev/changelog.md
@@ -7,6 +7,7 @@
 * Axis: Improved alignment of 90º rotated ticks (#1194, #1201) _Thanks @gigios_
 * Controls: Fix bug where middle-click-drag zoom rectangle would persist if combined with scroll wheel events (#1226) _Thanks @Elgot_
 * Scatter Plot: Fixed bug affecting plots where `YError` is set but `XError` is not (#1237, #1238) _Thanks @simmdan_
+* Palette: Added `Microcharts` colorset (#1235) _Thanks @arthurits_
 
 ## ScottPlot 4.1.17
 * Improved `RadarPlot.Update()` default arguments (#1097) _Thanks @arthurits_